feat: add shared logging module
